backend/requirements.txt
# FastAPI & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0

# Database Clients
motor==3.3.2
pymongo==4.6.1
qdrant-client==1.7.0
neo4j==5.16.0

# Celery & Redis
celery==5.3.6
redis==5.0.1

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
beautifulsoup4==4.12.3
httpx==0.26.0
trafilatura==1.6.4

# ML & Embeddings
sentence-transformers==2.2.2
spacy==3.7.2
openai==1.10.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
tenacity==8.2.3
structlog==24.1.0
